Refactor: move composables out of fragment
